requirements.txt
conllu
ufal.udpipe
pandas
requests
